Add DNS-PERSIST
